feat(autopsy): add keyword search panel
